@@ -1,18 +1,17 @@
 ﻿using System.Windows.Controls;
 
-namespace LiveSense.Common.Controls
-{
-    /// <summary>
-    /// Interaction logic for ErrorMessageDialog.xaml
-    /// </summary>
-    public partial class ErrorMessageDialog : UserControl
-    {
-        public string Message { get; set; }
+namespace LiveSense.Common.Controls;
 
-        public ErrorMessageDialog(string message)
-        {
-            InitializeComponent();
-            Message = message;
-        }
+/// <summary>
+/// Interaction logic for ErrorMessageDialog.xaml
+/// </summary>
+public partial class ErrorMessageDialog : UserControl
+{
+    public string Message { get; set; }
+
+    public ErrorMessageDialog(string message)
+    {
+        InitializeComponent();
+        Message = message;
     }
 }
