@@ -4,38 +4,38 @@ using System.Reflection;
 using System.Windows.Controls;
 using System.Windows.Navigation;
 
-namespace LiveSense.Common.Controls
+namespace LiveSense.Common.Controls;
+
+/// <summary>
+/// Interaction logic for InformationMessageDialog.xaml
+/// </summary>
+public partial class InformationMessageDialog : UserControl
 {
-    /// <summary>
-    /// Interaction logic for InformationMessageDialog.xaml
-    /// </summary>
-    public partial class InformationMessageDialog : UserControl
+    private readonly bool _showCheckbox;
+
+    public string VersionText => $"v{Assembly.GetEntryAssembly().GetName().Version}";
+    public bool ShowCheckbox => _showCheckbox;
+    public bool DontShowAgain { get; set; }
+
+    public InformationMessageDialog(bool showCheckbox)
     {
-        private readonly bool _showCheckbox;
+        _showCheckbox = showCheckbox;
 
-        public string VersionText => $"v{Assembly.GetEntryAssembly().GetName().Version}";
-        public bool ShowCheckbox => _showCheckbox;
-        public bool DontShowAgain { get; set; }
+        InitializeComponent();
+    }
 
-        public InformationMessageDialog(bool showCheckbox)
+    public void OnDismiss()
+    {
+        DialogHost.CloseDialogCommand.Execute(ShowCheckbox ? DontShowAgain : null, null);
+    }
+
+    public void OnNavigate(object sender, RequestNavigateEventArgs e)
+    {
+        Process.Start(new ProcessStartInfo()
         {
-            _showCheckbox = showCheckbox;
-
-            InitializeComponent();
-        }
-
-        public void OnDismiss()
-        {
-            DialogHost.CloseDialogCommand.Execute(ShowCheckbox ? DontShowAgain : null, null);
-        }
-
-        public void OnNavigate(object sender, RequestNavigateEventArgs e)
-        {
-            Process.Start(new ProcessStartInfo() {
-                FileName = e.Uri.AbsoluteUri,
-                UseShellExecute = true
-            });
-            e.Handled = true;
-        }
+            FileName = e.Uri.AbsoluteUri,
+            UseShellExecute = true
+        });
+        e.Handled = true;
     }
 }
