@@ -1,16 +1,14 @@
-﻿using System;
-using System.Globalization;
+﻿using System.Globalization;
 using System.Net;
 using System.Windows.Data;
 
-namespace LiveSense.Common.Converters
-{
-    public class IPEndPointToStringConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-            => value is IPEndPoint endpoint ? endpoint.ToString() : null;
+namespace LiveSense.Common.Converters;
 
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-            => value is string s && IPEndPoint.TryParse(s, out var endpoint) ? endpoint : null;
-    }
+public class IPEndPointToStringConverter : IValueConverter
+{
+    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        => value is IPEndPoint endpoint ? endpoint.ToString() : null;
+
+    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        => value is string s && IPEndPoint.TryParse(s, out var endpoint) ? endpoint : null;
 }
