@@ -2,47 +2,45 @@
 using LiveSense.Common.Messages;
 using Newtonsoft.Json.Linq;
 using Stylet;
-using System;
 
-namespace LiveSense.MotionSource
+namespace LiveSense.MotionSource;
+
+public abstract class AbstractMotionSource : Screen, IMotionSource, IDisposable, IHandle<AppSettingsMessage>
 {
-    public abstract class AbstractMotionSource : Screen, IMotionSource, IDisposable, IHandle<AppSettingsMessage>
+    public abstract string Name { get; }
+
+    protected AbstractMotionSource(IEventAggregator eventAggregator)
     {
-        public abstract string Name { get; }
+        eventAggregator.Subscribe(this);
+    }
 
-        protected AbstractMotionSource(IEventAggregator eventAggregator)
+    public abstract float GetValue(DeviceAxis axis);
+
+    protected abstract void HandleSettings(JObject settings, AppSettingsMessageType type);
+    public void Handle(AppSettingsMessage message)
+    {
+        if (message.Type == AppSettingsMessageType.Saving)
         {
-            eventAggregator.Subscribe(this);
+            if (!message.Settings.EnsureContainsObjects("MotionSource", Name)
+             || !message.Settings.TryGetObject(out var settings, "MotionSource", Name))
+                return;
+
+            HandleSettings(settings, message.Type);
         }
-
-        public abstract float GetValue(DeviceAxis axis);
-
-        protected abstract void HandleSettings(JObject settings, AppSettingsMessageType type);
-        public void Handle(AppSettingsMessage message)
+        else if (message.Type == AppSettingsMessageType.Loading)
         {
-            if (message.Type == AppSettingsMessageType.Saving)
-            {
-                if (!message.Settings.EnsureContainsObjects("MotionSource", Name)
-                 || !message.Settings.TryGetObject(out var settings, "MotionSource", Name))
-                    return;
+            if (!message.Settings.TryGetObject(out var settings, "MotionSource", Name))
+                return;
 
-                HandleSettings(settings, message.Type);
-            }
-            else if (message.Type == AppSettingsMessageType.Loading)
-            {
-                if (!message.Settings.TryGetObject(out var settings, "MotionSource", Name))
-                    return;
-
-                HandleSettings(settings, message.Type);
-            }
+            HandleSettings(settings, message.Type);
         }
+    }
 
-        protected virtual void Dispose(bool disposing) { }
+    protected virtual void Dispose(bool disposing) { }
 
-        public void Dispose()
-        {
-            Dispose(disposing: true);
-            GC.SuppressFinalize(this);
-        }
+    public void Dispose()
+    {
+        Dispose(disposing: true);
+        GC.SuppressFinalize(this);
     }
 }
