@@ -1,18 +1,17 @@
 ﻿using Newtonsoft.Json;
 using Stylet;
 
-namespace LiveSense.OutputTarget
-{
-    [JsonObject(MemberSerialization.OptIn)]
-    public class DeviceAxisSettings : PropertyChangedBase
-    {
-        [JsonProperty] public int Minimum { get; set; }
-        [JsonProperty] public int Maximum { get; set; }
+namespace LiveSense.OutputTarget;
 
-        public DeviceAxisSettings()
-        {
-            Minimum = 0;
-            Maximum = 100;
-        }
+[JsonObject(MemberSerialization.OptIn)]
+public class DeviceAxisSettings : PropertyChangedBase
+{
+    [JsonProperty] public int Minimum { get; set; }
+    [JsonProperty] public int Maximum { get; set; }
+
+    public DeviceAxisSettings()
+    {
+        Minimum = 0;
+        Maximum = 100;
     }
 }
