@@ -1,20 +1,19 @@
 ﻿using Stylet;
 
-namespace LiveSense.Service
+namespace LiveSense.Service;
+
+public class ServiceTip : PropertyChangedBase
 {
-    public class ServiceTip : PropertyChangedBase
+    public string Service { get; }
+    public string Username { get; }
+    public int Amount { get; }
+
+    public float? Progress { get; set; }
+
+    public ServiceTip(string service, string username, int amount)
     {
-        public string Service { get; }
-        public string Username { get; }
-        public int Amount { get; }
-
-        public float? Progress { get; set; }
-
-        public ServiceTip(string service, string username, int amount)
-        {
-            Service = service;
-            Username = username;
-            Amount = amount;
-        }
+        Service = service;
+        Username = username;
+        Amount = amount;
     }
 }
