@@ -6,110 +6,108 @@ using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Linq;
 using Stylet;
 using StyletIoC;
-using System;
 using System.Diagnostics;
 using System.IO;
 using System.Windows;
 using System.Windows.Input;
 
-namespace LiveSense.ViewModels
+namespace LiveSense.ViewModels;
+
+public class RootViewModel : Conductor<IScreen>.Collection.AllActive
 {
-    public class RootViewModel : Conductor<IScreen>.Collection.AllActive
+    private readonly IEventAggregator _eventAggregator;
+
+    [Inject] public MotionSourceViewModel MotionSource { get; internal set; }
+    [Inject] public TipQueueViewModel TipQueue { get; internal set; }
+    [Inject] public OutputTargetViewModel OutputTarget { get; set; }
+    [Inject] public ServiceViewModel Service { get; set; }
+
+    public RootViewModel(IEventAggregator eventAggregator)
     {
-        private readonly IEventAggregator _eventAggregator;
+        _eventAggregator = eventAggregator;
 
-        [Inject] public MotionSourceViewModel MotionSource { get; internal set; }
-        [Inject] public TipQueueViewModel TipQueue { get; internal set; }
-        [Inject] public OutputTargetViewModel OutputTarget { get; set; }
-        [Inject] public ServiceViewModel Service { get; set; }
-
-        public RootViewModel(IEventAggregator eventAggregator)
+        JsonConvert.DefaultSettings = () =>
         {
-            _eventAggregator = eventAggregator;
-
-            JsonConvert.DefaultSettings = () =>
+            var settings = new JsonSerializerSettings
             {
-                var settings = new JsonSerializerSettings
-                {
-                    Formatting = Formatting.Indented
-                };
-
-                settings.Converters.Add(new StringEnumConverter());
-                return settings;
+                Formatting = Formatting.Indented
             };
-        }
 
-        protected override void OnActivate()
-        {
-            Items.Add(MotionSource);
-            Items.Add(TipQueue);
-            Items.Add(OutputTarget);
-            Items.Add(Service);
+            settings.Converters.Add(new StringEnumConverter());
+            return settings;
+        };
+    }
 
-            ActivateAndSetParent(Items);
-            base.OnActivate();
-        }
+    protected override void OnActivate()
+    {
+        Items.Add(MotionSource);
+        Items.Add(TipQueue);
+        Items.Add(OutputTarget);
+        Items.Add(Service);
 
-        public void OnInformationClick()
-            => _ = Execute.OnUIThreadAsync(() => DialogHost.Show(new InformationMessageDialog(showCheckbox: false), "RootDialog"));
+        ActivateAndSetParent(Items);
+        base.OnActivate();
+    }
 
-        public void OnLoaded(object sender, EventArgs e)
-        {
-            Execute.PostToUIThread(async () =>
-            {
-                var settings = ReadSettings();
-                _eventAggregator.Publish(new AppSettingsMessage(settings, AppSettingsMessageType.Loading));
+    public void OnInformationClick()
+        => _ = Execute.OnUIThreadAsync(() => DialogHost.Show(new InformationMessageDialog(showCheckbox: false), "RootDialog"));
 
-                if (!settings.TryGetValue("DisablePopup", out var disablePopupToken) || !disablePopupToken.Value<bool>())
-                {
-                    var result = await DialogHost.Show(new InformationMessageDialog(showCheckbox: true), "RootDialog").ConfigureAwait(true);
-                    if (result is not bool disablePopup || !disablePopup)
-                        return;
-
-                    settings["DisablePopup"] = true;
-                    WriteSettings(settings);
-                }
-            });
-        }
-
-        public void OnClosing(object sender, EventArgs e)
+    public void OnLoaded(object sender, EventArgs e)
+    {
+        Execute.PostToUIThread(async () =>
         {
             var settings = ReadSettings();
-            _eventAggregator.Publish(new AppSettingsMessage(settings, AppSettingsMessageType.Saving));
-            WriteSettings(settings);
-        }
+            _eventAggregator.Publish(new AppSettingsMessage(settings, AppSettingsMessageType.Loading));
 
-        private JObject ReadSettings()
-        {
-            var path = Path.Join(Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName), "LiveSense.config.json");
-            if (!File.Exists(path))
-                return new JObject();
-
-            try
+            if (!settings.TryGetValue("DisablePopup", out var disablePopupToken) || !disablePopupToken.Value<bool>())
             {
-                return JObject.Parse(File.ReadAllText(path));
+                var result = await DialogHost.Show(new InformationMessageDialog(showCheckbox: true), "RootDialog").ConfigureAwait(true);
+                if (result is not bool disablePopup || !disablePopup)
+                    return;
+
+                settings["DisablePopup"] = true;
+                WriteSettings(settings);
             }
-            catch (JsonException)
-            {
-                return new JObject();
-            }
-        }
+        });
+    }
 
-        private void WriteSettings(JObject settings)
+    public void OnClosing(object sender, EventArgs e)
+    {
+        var settings = ReadSettings();
+        _eventAggregator.Publish(new AppSettingsMessage(settings, AppSettingsMessageType.Saving));
+        WriteSettings(settings);
+    }
+
+    private JObject ReadSettings()
+    {
+        var path = Path.Join(Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName), "LiveSense.config.json");
+        if (!File.Exists(path))
+            return new JObject();
+
+        try
         {
-            var path = Path.Join(Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName), "LiveSense.config.json");
-            File.WriteAllText(path, settings.ToString());
+            return JObject.Parse(File.ReadAllText(path));
         }
-
-        public void OnMouseDown(object sender, MouseButtonEventArgs e)
+        catch (JsonException)
         {
-            if (sender is not Window window)
-                return;
-
-            if (e.LeftButton != MouseButtonState.Pressed)
-                return;
-
-            window.DragMove();
+            return new JObject();
         }
+    }
+
+    private void WriteSettings(JObject settings)
+    {
+        var path = Path.Join(Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName), "LiveSense.config.json");
+        File.WriteAllText(path, settings.ToString());
+    }
+
+    public void OnMouseDown(object sender, MouseButtonEventArgs e)
+    {
+        if (sender is not Window window)
+            return;
+
+        if (e.LeftButton != MouseButtonState.Pressed)
+            return;
+
+        window.DragMove();
     }
 }
